requirements.txt
feedparser==6.0.10
python-dotenv==1.0.0
pymongo==4.6.1
requests==2.31.0
azure-search-documents==11.4.0
sentence-transformers==2.2.2
